requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-filter==2.2.0
djangorestframework==3.11.0
Faker==4.0.2
Markdown==3.2.1
mixer==6.1.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
